MGMT-1739 - prepare service config file that will include info extracted from OCP release image needed by other services
