+++ Dockerfile.ignition-manifests-and-kubeconfig-generate
@@ -39,6 +39,7 @@ RUN cp $TAR_DIR/oc $WORK_DIR/oc
 COPY ./render_files.py $WORK_DIR
 COPY ./utils.py $WORK_DIR
 COPY ./bmh_utils.py $WORK_DIR
+COPY ./oc_utils.py $WORK_DIR
 COPY ./test_utils.py $WORK_DIR
 
 ENV WORK_DIR=$WORK_DIR
